The Comment model includes the following fields:
bookId: The ID of the associated book the comment belongs to. It is a required field and references the 'Book' model.
content: The content of the comment. It is a required field and stored as a string.
The Comment model is exported as a Mongoose model named 'Comment' to be used for CRUD operations on the comments collection.